Improve error-reporting.
Rather than showing an error when we process a URL, then
terminating, we now save the errors we encounter when processing
and report them en masse once we've finished all the feeds.

This means a broken feed in the middle of our processing loop
will not prevent other feeds from being processed.

This closes #9.