Call sendmail with -i option
By default, sendmail stops reading input when it encounters a line
containing a single '.' character.  This is rare in normal text,
but can happen in a quoted-printable encoding when a line is broken
just before the terminating '.'.  I saw it happen on a random rss
feed item. The -i option causes sendmail to process such lines
normally