Retain comments in the feeds file
We assume that all comment and blank lines in the feeds file are
associated with the feed url that follows them.  We maintain
those comments and list or delete them when the url is listed
or deleted.  When a new feed url is added, a comment is added
that consists of the feed's Title.

These default Title comments can be created for a current feedlist
by the following command:
    cd ~/.rss2email && mv feeds feeds.tmp &&
        rss2email add $(grep '^[^#]' feeds.tmp) && rm feeds.tmp