feedlist: validate each feed url when its added
Currently, any random, non-url, string submitted via "rss2email add"
will be accepted and added to the feedlist.  This will result in
errors when "rss2email cron" attempts to retrieve the bogus feed.

This commit validates each feed url when it is added or imported
and invalid feed urls will not be added to the feedlist.