Allow excluding-in-reverse; mandatory regexp for sending
This pull-request closes #63, by allowing the support for the reverse
of excluding entries.

The following setting ensures that any entry in a feed containing the
term `foo` is not emailed:

        - exclude: foo

As a result of this pull-request we now support:

        http://example.com/feed/
          - include: bar

This will only email entries from the given feed that match the
regular expression bar.