Allow setting a per-feed HTTP User-Agent
This pull-request adds a new "user-agent" configuration setting,
which allows a user-agent to be specified upon a per-feed basis.

We've always had a hardwired version, which was added as a result
of some sites failing to return content if the header was missing
(e.g. reddit, which doesn't like wget/curl agents, or empty values).

Now, as a result of DreamWidth making changes we've been forced to
allow this to be configured.