Avoid pruning state if we see errors.
This pull-request updates the processing of the feeds to avoid
pruning our history of recently-seen-state if we received any
errors fetching our feeds.

This is suboptimal, because the local state will continue to grow
if there are broken feeds in our configuration file.  However this
is better than constantly renotifying on previously-seen items
when/if transient errors occur.

To solve this properly would require a per-feed directory to maintain
state of entries beneath.  We will evolve in that direction but it
will require a migration plan first.

This closes #80.