Use BoltDB to store feed-state
This pull-request largely neuters the use of ~/.rss2email/state,
instead storing state of the feed-items inside a BoltDB database.

The migration here is seamless:

* When a feed is fetched we get a list of items.
* We look to see if the item is new, in the legacy database OR the new
  one.
  * If the item is new we record it ONLY in boltdb.
  * THen we remove the legacy state-file.

The end result is that legacy state-files will be cleaned up,
and removed and all state will end up in BoltDB.

The BoltDB file will continue to grow, as that is not yet pruned,
but that will be an easy change to make in the future.

As part of this we've also disabled CGO, and generated binaries
for all platforms.

This would close #82, when merged.