121 logging improvements (#122)
This pull-request closes #121 by adding a unified and consistent logging pattern:

Some usage examples are showing by `fmt.Printf` as are the errors at the end of a run - which will be reconsidered in the near future, but that aside **all** output is handled via a logger:

* There are three logging-levels and their output is defined:
   * DEBUG: For developers, bug-reporters, or the curious.
   * WARN: Showing "stuff that is important".
   * ERROR: For fatal errors.

The `-verbose` flag for the daemon/cron command basically just bumps the default log-level from WARN to DEBUG, to increase verbosity indirectly.  

To ease parsing/processing the logging uses the new `slog` package, and is "structured".  To make this even more structured it is possible to output logs in JSON format, via the use of an environmental variable.