Allow logging to a file (#123)
In #122 we updated our codebase to ensure that we used a common
logging handle to send our output messages to the user:

* This included warnings and errors by default.
* However "developer", or other internal messages, were also available.

This pull-request builds upon that work to write the log messages to
a file, as well as showing them to the user.  By default a file
`rss2email.log` is generated when the application starts and log
entries are appended to it.

However:

* LOG_FILE_PATH can be set to a different path.
* LOG_FILE_DISABLE can be used to disable this duplication.

Why do this?  Partly for reference, but also partly to allow a local
setup to view errors.

We could now resolve #119 by adding a cronjob:

        #!/bin/sh
        # show logs, and delete them, run once per day
        cat .../rss2email.log
        rm .../rss2email.log

(Of course this might not work 100% as the current approach assumes
the file is open forever in the case of `daemon` sub-command.)

This isn't a complete solution, but without getting into the whole
template-customization and more complex considerations it's not a
terrible thing to do.