Implement support for cached-fetches
This pull-request, once complete, will close #131 by avoiding the
need to process remote feeds which have not changed since their last
fetch.

We do this by retrieving "Etag" and "Last-Modified" headers when a
URL (i.e. feed) is fetched the first time, and use those values in
subsequent fetches via headers ("If-Non-Matche", and "If-Modified-Since").

If the remote feed has not changed we should get a 3xx HTTP-status
code, and the remote server should be greatful that it has less work
to do.