fix: Persist the HTTP-headers, used for caching, to disk
In the past we tried to save the HTTP-headers which we returned to
an in-memory cache (i.e. map), so that on all HTTP-requests after
the first we could use them to minimize load on our remote feeds.

However we did not _persist_ the headers, so that meant on startup
we'd always make one unconditional HTTP-fetch against each configured
feed.

This pull-request saves the values we've received and uses them on
startup, and future updates.

This should close #140.